requirements.txt
Pillow
gpiozero
lgpio
spidev
flask